Game start
A world that can spin with arrow keys or buttons. You may need to attach the world game object to the buttons for them to work on your system. Also not sure if my git ignore file is in the ""root" of the project directory," so there may be extra files.